Uniformity. Premature optimization. Simplify.
I had already started swapping out the C-style string formatting options with
``str.format()`` I finished that to make the code more uniform.

``_negotiate_value()`` used ``re.compile`` every time it was called, which is
expensive. I "cached" the regular expression object so it should offer a small
improvement to those using this excessively.

Finally, in ``HTTPKerberosAuth.__init__`` there was a check if kerberos was
None before trying to use it. That is unnecessary since this is now a separate
module.