Setup.py breaks if kerberos is not already installed.
When trying to install requests-kerberos setup.py cannot import
requets_kerberos.  Doing this causes a problem where requirements.txt
causes kerberos to be downloaded.  However, it is not yet installed
when requests-kerberos's setup.py is evaluated.  Since setup.py
imports requests_kerberos, and requests_kerberos import kerberos, the
installation fails.

See https://gist.github.com/pcn/5209734 for an example.

This commit fixes the problem by using the compiler module to parse
__init__.py, but doesn't evaluate it.  It then searches for the node
whose name is __version__ and uses the constant that would be assigned
to it for setup.py.

It does mean that version has to be a constant.  I don't think that'd
be a problem.