Fixing assertions regarding mocks not being called; resetting mocks
We need to reset mocks between tests if we want to be sure when we assert
whether or not they've been called and what they've been called with because if
we don't, they may have been called during other tests.

We mock.assert_not_called() isn't a thing, but it always returns a Truthy value
because it's a mock, and any method not defined will actually exist and return a
mock which is truthy. My head hurts. We need to assert that the special .called
parameter is False.