Fix hangs on streaming uploads
After the server responds with a "401 Unauthorized",
requests_kerberos.HTTPKerberosAuth.handle_response() gets called and
it resends the request with "Authorization" header.

When doing streaming uploads, however, this gets stuck because the
body data (a file-like object) is already consumed at the first
attempt of sending the request and empty.

To fix this, rewind the file-like object to where it was before the
initial send call.