Resolve potential concurrency issue
When used from different threads (or potentially as part of an event
loop), it is possible that two different requests to the same host
occurring concurrently could stomp on one anothers kerberos context in
the context dictionary. This could result in client or server (mutual)
authentication failures as the wrong context would be used at different
times.

To resolve this, we do two things:

 1. Within a call to `generate_request_header`, we bind the context we
    generate to a name local to the function, and always use the context
    via that name. As a result, any concurrently executing call to
    `generate_request_header` will be unable to rebind a new context to
    that name.
 2. When populating the context dictionary, we index by
    `PreparedRequest` object instead of request urls `hostname`. This is
    possible because when performing mutual authentication on a
    response, the `PreparedRequest` object which resulted in the
    response being processed is available as the `request` property on
    the response.

Unfortunately, this breaks our WinRM support because it would require
that the user provide the prepared request to the `winrm_wrap` and
`winrm_unwrap` functions. Since this is is intended to be a module
facilitating Kerberos/GSSAPI authentication for HTTP, we drop explicit
WinRM support, and instead attach the established context to the
response we return so that callers can implement that application
specific logic themselves.

Note: In addition to removing `winrm_wrap` and `winrm_unwrap`, this
changes the type signature of `generate_request_header`, so that could
be considered a breaking change. We may want to prefix some of these
intended to be internal-use functions with underscores as a warning to
users.