Try winkerberos first
In the previous order, the `ImportError` is masked on Linux and leads to a confusing message about not finding the `winkerberos` which just confuses most Linux users.

This ordering lets the real issue be exposed for Linux users. Windows users are less likely to be missing pieces and there should be fewer users of this software on Windows too.