Avoid memory leak
We should not store the request or the established kerberos context on
the authentication handler forever. Once it's no longer necessary (we've
authenticated the server and attached the context to the response we're
returning to the user for them to do with as they wish), we should drop
it so it can potentially be collected.